DTS-48150:Refactor ScmPRSizeServiceImpl
